refactor: move message merging utilities to shared module
