+++ packages/shared/src/index.ts
@@ -3,3 +3,4 @@ export * from './warn'
 export * from './error'
 export * from './emittable'
 export * from './emitter'
+export * from './messages'
